Csr 11 skill acquire and update rework (#14)
* CSR-11 - acquire skills after levels interval and toogle unlimited stats

* CSR-11 - changing how skills are currently upgraded (now new skill after 3 upgrades).

* CSR-11 - fixing skills unlimited upgrades error caused by byte type

+++ Assets/Scripts/Projectiles/Projectile.cs
@@ -14,7 +14,7 @@ namespace Assets.Scripts.Projectiles
     {
         [SerializeField] private ProjectileConfigSO _config;
         [SerializeField] private SphereCollider _sphereCollider;
-        private byte _piercedCounter;
+        private int _piercedCounter;
         private bool _isInitialized;
         private bool _isAlive = true;
